Split out getSitePost
